Appstore.html page addition to assignment
(Sorry if it is a major pain to change this into .jsp. it is a pain for me with the testing environment that I use. I spent more time than I thought I would, but then again, it could take more too. Add to it if you desire. I think I covered our basis. You may need to add a link to the "login" jsp like you have with the others, that's where my "get" button leads to)